added startup script
