+++ package.json
@@ -3,9 +3,9 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@baltimorecounty/dotgov-components": "^1.0.3",
+    "@baltimorecounty/dotgov-components": "^1.0.19",
     "@baltimorecounty/javascript-utilities": "^1.0.7",
-    "@baltimorecounty/react-filter-list": "^0.2.2",
+    "@baltimorecounty/react-filter-list": "1.0.4",
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.3.2",
     "@testing-library/user-event": "^7.1.2",
@@ -13,6 +13,7 @@
     "react": "^16.13.1",
     "react-app-polyfill": "^1.0.6",
     "react-dom": "^16.13.1",
+    "react-router-dom": "^5.2.0",
     "react-scripts": "3.4.1"
   },
   "scripts": {
